clean BuildFiles in EventFilter
