+++ EventFilter/L1TXRawToDigi/BuildFile.xml
@@ -1,0 +1,1 @@
+<use name="FWCore/MessageLogger"/>
